Refactor Inspector Controls out of the Query Loop block
Add an attribute `disabledInspectorControls` to let users decide which
controls should be displayed.
