+++ packages/block-library/src/query/block.json
@@ -37,6 +37,9 @@
 			"default": {
 				"type": "list"
 			}
+		},
+		"disabledInspectorControls": {
+			"type": "array"
 		}
 	},
 	"providesContext": {
